fix: update apache commons and support older api
In 8e41347512f0ecc374617f970990e31609464c39 we changed compression
libraries in order to increase Android minimum compatibility from
SDK 26 (Android 8) to SDK 22 (Android 5.1)

With latest Flagship app changes we encountered build issues due to
this change. This is mainly due to removal of geolocation libraries
that shared common dependencies

We should be able to fix those dependencies, but because latest
ReactNative versions break support for Android 5 and 6 and because we
will also rework this library for nitro-modules support, we chose not
to investigate this for now and revert this commit instead. However we
also loose support for Android 7 but this impact only a very small
number of users (less than 10)

This commit is a revert from 8e41347512f0ecc374617f970990e31609464c39